fix(windows): zero Windows clippy warnings, fix test-only path and sender lookup
- gate/prune cfg(unix)-only imports, dead code, and fix style lints in
  Windows-only code paths (windows clippy now passes -D warnings clean)
- test temp-file helpers used thread::current().name() (contains '::',
  illegal in Windows filenames, os error 123) — sanitize to '_'
- extract_agent_signal_sender on Windows also searches the parent dir
  of the test binary (target/release) and CARGO_BIN_EXE_* fallback

+++ src/platform/process.rs
@@ -37,8 +37,8 @@ pub fn spawn_detached(command: &mut Command) -> io::Result<Child> {
         .stdout(Stdio::null())
         .stderr(Stdio::null())
         .creation_flags(
-            windows_sys::Win32::System::Threading::CREATE_NEW_PROCESS_GROUP as u32
-                | windows_sys::Win32::System::Threading::DETACHED_PROCESS as u32,
+            windows_sys::Win32::System::Threading::CREATE_NEW_PROCESS_GROUP
+                | windows_sys::Win32::System::Threading::DETACHED_PROCESS,
         )
         .spawn()
 }
